pyproject.toml
[project]
name = "data-for-tsfms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "chronos-forecasting>=2.2.2",
    "datasets>=4.6.1",
    "gluonts[arrow]>=0.16.2",
    "mlflow>=3.10.0",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "typer>=0.24.1",
    "typer-config>=1.5.1",
]
